Rendering tests on interbit-template
